+++ CHANGELOG.md
@@ -9,8 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `JsonPatchOperations` - All possible JSON patch operations
+- `JsonPatch` - A valid JSON patch
+
+### Fixed
+
 - `JsonValue` in favor of `JSONValue`. Casing should be about word spacing only.
   Using all capitals for an abbreviation makes names harder to read.
+- `isArray(x)` - Is x an array?
+- `isObject(x)` - Is x an object?
 
 ### Deprecated